+++ http-client/ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.4.18
+
+* Deprecate closeManager [#136](https://github.com/snoyberg/http-client/issues/136) [#137](https://github.com/snoyberg/http-client/issues/137)
+
 ## 0.4.17
 
 * Case insensitive proxy environment variables [#135](https://github.com/snoyberg/http-client/issues/135)