+++ http-conduit/ChangeLog.md
@@ -1,3 +1,7 @@
+## 2.1.7
+
+* Deprecate `conduitManagerSettings`, re-export `tlsManagerSettings` [#136](https://github.com/snoyberg/http-client/issues/136) [#137](https://github.com/snoyberg/http-client/issues/137)
+
 ## 2.1.6
 
 * Deprecate `withManager` and `withManagerSettings`